Add a CogniteLogger abstract class
